clean up

+++ core/src/fs/image.rs
@@ -1,6 +1,5 @@
 use image::{imageops, io::Reader, DynamicImage, EncodableLayout, GenericImageView};
 use rayon::prelude::{IntoParallelIterator, ParallelIterator};
-// use rocket::tokio::{self, task::JoinHandle};
 use std::{
 	fs::File,
 	io::{Read, Write},
@@ -80,11 +79,15 @@ pub fn generate_thumbnail(id: &str, path: &str) -> ProcessFileResult<PathBuf> {
 	let (_, buf) = media_file::get_page(path, 1)?;
 	let webp_buf = webp_from_bytes(&buf)?;
 
-	let thumbnail_path = get_thumbnails_dir().join(format!("{}.webp", id));
+	let thumbnail_path = get_thumbnails_dir().join(format!("{}.webp", &id));
 
-	let mut webp_image = File::create(&thumbnail_path)?;
+	if !thumbnail_path.exists() {
+		let mut webp_image = File::create(&thumbnail_path)?;
 
-	webp_image.write_all(&webp_buf)?;
+		webp_image.write_all(&webp_buf)?;
+	} else {
+		log::trace!("Thumbnail already exists for {}", &id);
+	}
 
 	Ok(thumbnail_path)
 }
@@ -107,29 +110,6 @@ pub fn generate_thumbnails(media: Vec<media::Data>) -> ProcessFileResult<Vec<Pat
 			}
 		})
 		.collect::<Vec<PathBuf>>();
-
-	// let tasks: Vec<JoinHandle<ProcessFileResult<PathBuf>>> = media
-	// 	.into_iter()
-	// 	.map(|m| {
-	// 		tokio::task::spawn_blocking(move || {
-	// 			generate_thumbnail(m.id.as_str(), &m.path.as_str())
-	// 		})
-	// 	})
-	// 	.collect();
-
-	// let results = futures::future::join_all(tasks)
-	// 	.await
-	// 	.into_iter()
-	// 	.filter_map(|res| res.ok())
-	// 	.filter_map(|res| {
-	// if res.is_err() {
-	// 	log::error!("Error generating thumbnail: {:?}", res.err());
-	// 	None
-	// } else {
-	// 	res.ok()
-	// }
-	// 	})
-	// 	.collect::<Vec<PathBuf>>();
 
 	log::debug!("Generated the following thumbnails: {:?}", results);
 
@@ -156,7 +136,7 @@ pub fn remove_thumbnail(id: &str) -> ProcessFileResult<()> {
 	Ok(())
 }
 
-pub fn remove_thumbnails(id_list: &[&str]) -> ProcessFileResult<()> {
+pub fn remove_thumbnails(id_list: &[String]) -> ProcessFileResult<()> {
 	for id in id_list {
 		// TODO: not sure I want the entire process to fail if one thumbnail fails to delete...
 		// for now, I will leave it as is. I can't see to many cases where this would happen, but
@@ -167,6 +147,6 @@ pub fn remove_thumbnails(id_list: &[&str]) -> ProcessFileResult<()> {
 	Ok(())
 }
 
-// TODO: tests
+// TODO: tests!
 #[cfg(test)]
 mod tests {}
